[proliferate] organize app and main hooks
