+++ desktop/src/App.tsx
@@ -10,9 +10,9 @@ import { MacWindowControlsSafeArea } from "@/components/ui/MacWindowControlsSafe
 import { SessionModelAvailabilityDialog } from "@/components/workspace/chat/launch/SessionModelAvailabilityDialog"
 import { applyAppearancePreference, initializeTheme } from "@/config/theme"
 import { APP_ROUTES, LEGACY_APP_ROUTES } from "@/config/app-routes"
-import { useAppCommandActions } from "@/hooks/app/use-app-command-actions"
-import { useExportRunningAgentCount } from "@/hooks/app/use-export-running-agent-count"
-import { useAppShortcuts } from "@/hooks/app/use-app-shortcuts"
+import { useExportRunningAgentCount } from "@/hooks/app/lifecycle/use-export-running-agent-count"
+import { useAppShortcuts } from "@/hooks/app/lifecycle/use-app-shortcuts"
+import { useAppCommandActions } from "@/hooks/app/workflows/use-app-command-actions"
 import { useAuthBootstrap } from "@/hooks/auth/use-auth-bootstrap"
 import { useAgentAutoReconcile } from "@/hooks/agents/lifecycle/use-agent-auto-reconcile"
 import { useLocalAutomationExecutor } from "@/hooks/automations/lifecycle/use-local-automation-executor"
