+++ desktop/src/hooks/main/lifecycle/use-main-screen-shortcuts.test.tsx
