+++ desktop/src/lib/domain/workspaces/creation/publish-dialog-state.ts
